make commands module, add error
